move useState and handleFavorite up to App component and pass favorite and handleFavorite as prop back down to Recipe component
