+++ client/src/App.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import "./App.css";
-
+import { useState } from "react";
 import Home from "../src/pages/Home";
 import NavBar from "./components/NavBar";
 import Recipes from "../src/pages/Recipes";
@@ -9,6 +9,13 @@ import useRecipes from "./hooks/useRecipes";
 
 function App() {
   const { recipes } = useRecipes();
+  const [favorite, setFavorite] = useState({});
+  const handleFavorite = (recipeId) => {
+    setFavorite((prevFavorite) => ({
+      ...prevFavorite,
+      [recipeId]: !prevFavorite[recipeId],
+    }));
+  };
   return (
     <>
       <Router>
@@ -16,7 +23,7 @@ function App() {
         <Routes>
           <Route path="/login" element={<Login />} />
           <Route path="/" element={<Home />} />
-          <Route path="/recipes" element={<Recipes recipes={recipes} />} />
+          <Route path="/recipes" element={<Recipes recipes={recipes} favorite={favorite} handleFavorite={handleFavorite}/>} />
         </Routes>
       </Router>
     </>
